Fix: Retira o elemento button que estava envolvendo o GoogleOAuth e transfere este componente da pasta SignUp para uma pasta propria.
